Tests/LibWeb/Text/input/canvas/basic.html

<script src="../include.js"></script>
<script>
    test(() => {
        let canvas = document.createElement("canvas");
        let context = canvas.getContext("2d");
        context.font = '20px DenLabSans';
        println(context.font);
        context.font = '!!!'; // Invalid value, should be ignored.
        println(context.font);
    });
</script>
