Tests/LibWeb/Text/expected/canvas/basic.txt

normal normal 20px DenLabSans
normal normal 20px DenLabSans
